+++ jungle-base/include/jungle/os/shared_memory.h
@@ -16,7 +16,7 @@ namespace jungle::os {
 struct alignas(cacheline_size) shm_header {
     usize size;
     std::atomic<usize> holder_count;
-    const u64 build_id;
+    const u128 build_id;
 };
 
 static_assert(
